feat(collect): added continuous sensor data collection and imu modality #1000
Implemented the ability to collect data in a continuous manner, i.e.,
collect all sensor messages and record them all in the dataset during
data collection. To do this, a new 'continuous' field in the configuration for a
given sensor was created and included in the template.

To maintain backwards compatibility with older datasets, if there is no
'continuous' field, it defaults to False and works as it did previously.

The 'continuous' field was also added to the configuration YAML
template.

Additionally, support for the IMU sensor modality in data collection was also added.